flytt AuthContext til egen fil
