+++ src/frontend/context/AppContext.tsx
@@ -8,6 +8,7 @@ import { HttpProvider, loggFeil, useHttp } from '@navikt/familie-http';
 import type { ISaksbehandler, Ressurs } from '@navikt/familie-typer';
 import { RessursStatus } from '@navikt/familie-typer';
 
+import { AuthProvider, useAuthContext } from './AuthContext';
 import type { IToast, ToastTyper } from '../komponenter/Toast/typer';
 import { BehandlerRolle } from '../typer/behandling';
 import type { IRestTilgang } from '../typer/person';
@@ -40,40 +41,6 @@ const initalState: IModal = {
 interface IProps extends PropsWithChildren {
     autentisertSaksbehandler: ISaksbehandler | undefined;
 }
-
-interface AuthProviderExports {
-    autentisert: boolean;
-    settAutentisert: (autentisert: boolean) => void;
-    innloggetSaksbehandler: ISaksbehandler | undefined;
-}
-
-const AuthContext = createContext<AuthProviderExports | undefined>(undefined);
-
-const AuthProvider = ({ autentisertSaksbehandler, children }: IProps) => {
-    const [autentisert, settAutentisert] = React.useState(true);
-    const [innloggetSaksbehandler, settInnloggetSaksbehandler] =
-        React.useState(autentisertSaksbehandler);
-
-    useEffect(() => {
-        if (autentisertSaksbehandler) {
-            settInnloggetSaksbehandler(autentisertSaksbehandler);
-        }
-    }, [autentisertSaksbehandler]);
-
-    return (
-        <AuthContext.Provider value={{ autentisert, settAutentisert, innloggetSaksbehandler }}>
-            {children}
-        </AuthContext.Provider>
-    );
-};
-
-const useAuthContext = () => {
-    const context = React.useContext(AuthContext);
-    if (!context) {
-        throw new Error('useAuth må brukes innenfor AuthProvider');
-    }
-    return context;
-};
 
 interface AppContentContextValue {
     autentisert: boolean;
